requirements.txt
art=5.7
psycopg2-binary=2.9.3